+++ Sources/VCore/Extensions/Foundation/Set.Toggle.swift
@@ -1,0 +1,28 @@
+//
+//  Set.Toggle.swift
+//  VCore
+//
+//  Created by Vakhtang Kontridze on 01.05.22.
+//
+
+import Foundation
+
+// MARK: - Toggle Set Element
+extension Set {
+    /// Inserts element in a set if it's not present, and if present, removes it.
+    ///
+    /// Usage Example
+    ///
+    ///     var nums: Set<Int> = [1, 3, 5]
+    ///     nums.toggle(1) // [3, 5]
+    ///     nums.toggle(1) // [1, 3, 5]
+    ///
+    public mutating func toggle(
+        _ element: Element
+    ) {
+        switch contains(element) {
+        case false: insert(element)
+        case true: remove(element)
+        }
+    }
+}